compozed/requirements.txt
requests
django
djangorestframework
channels
asgi_redis
